simulation/requirements.txt
pyprimo
numpy
panda
scikit_learn
lightgbm
seaborn
matplotlib


